+++ src/moviegraphbenchmark/create_graph.py
@@ -1,13 +1,13 @@
 import ast
 import logging
 import os
-import zipfile
 from typing import List, Set, Tuple, Optional
 
 import click
 
 from moviegraphbenchmark.get_imdb_data import download_if_needed
-from moviegraphbenchmark.utils import download_file
+from moviegraphbenchmark.utils import download_github_folder
+import moviegraphbenchmark
 
 DTYPE_DOUBLE = "<http://www.w3.org/2001/XMLSchema#double>"
 DTYPE_NON_NEG_INT = "<http://www.w3.org/2001/XMLSchema#nonNegativeInteger>"
@@ -387,19 +387,6 @@ def write_files(
             out_writer_rel.write("\t".join(t) + "\n")
 
 
-def _download_data(data_path: str):
-    if not os.path.exists(data_path):
-        os.makedirs(data_path)
-    download_file(
-        "https://cloud.scadsai.uni-leipzig.de/index.php/s/4WwwG7yHWPAiNnc/download/ScadsMovieGraphBenchmark.zip",
-        data_path,
-    )
-    zip_path = os.path.join(data_path, "ScadsMovieGraphBenchmark.zip")
-    with zipfile.ZipFile(zip_path, "r") as zip_ref:
-        zip_ref.extractall(data_path)
-    os.remove(zip_path)
-
-
 def _create_data_path() -> Tuple[str, bool]:
     file_path = os.path.abspath(__file__)
     repo_path = os.path.split(os.path.split(os.path.split(file_path)[0])[0])[0]
@@ -427,14 +414,16 @@ def _create_graph_data(data_path: Optional[str] = None) -> str:
     existing_data_path = False
     if data_path is None:
         data_path, existing_data_path = _create_data_path()
-    print(data_path)
     # check if data was already created
     if os.path.exists(os.path.join(data_path, "imdb-tmdb", "rel_triples_1")):
-        logger.info(f"Data already present in {data_path}")
-        return data_path
+        if not os.path.exists(os.path.join(data_path, "imdb_intra_ent_links")):
+            logger.info(f"Old data already present in {data_path}, will update...")
+        else:
+            logger.info(f"Data already present in {data_path}")
+            return data_path
     logger.info(f"Using data path: {data_path}")
-    if not existing_data_path or not os.path.exists(os.path.join(data_path, "imdb-tmdb", "rel_triples_2")):
-        _download_data(data_path)
+    if not os.path.exists(os.path.join(data_path, "imdb_intra_ent_links")):
+        download_github_folder(data_path, moviegraphbenchmark.__version__)
     imdb_path = os.path.join(data_path, "imdb")
     download_if_needed(imdb_path)
     allowed = get_allowed(os.path.join(data_path, "imdb", "allowed"))