+++ CHANGELOG.md
@@ -7,9 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2024-03-13
+
+### Fixed
+
+- Remove intra-dataset links from ent_links
+
+### Added
+
+- intra_ent_links to ERDataset
+- cluster files
+- multi_source_cluster file
+
 ## [1.0.2] - 2023-03-24
 
-### Changed 
+### Changed
 
 - Loosen dependency restrictions to allow every Python >=3.7.1
 
@@ -24,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Make packagable and enable loading of datasets
 
+[1.1.0]: https://github.com/ScaDS/MovieGraphBenchmark/releases/tag/v1.0.2
 [1.0.2]: https://github.com/ScaDS/MovieGraphBenchmark/releases/tag/v1.0.2
 [1.0.1]: https://github.com/ScaDS/MovieGraphBenchmark/releases/tag/v1.0.1
 [1.0.0]: https://github.com/ScaDS/MovieGraphBenchmark/releases/tag/v1.0.0